+++ deploy/kolla/src/palimpsest_kolla/__init__.py
@@ -1,1 +1,0 @@
-"""Palimpsest Kolla-Ansible packaging marker package."""